requirements.txt
# Core SDK
claude-agent-sdk>=0.1.0

# Google ADK for orchestration
google-adk>=0.1.0
google-generativeai>=0.8.0

# Gmail integration
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.110.0

# Voice streaming
websockets>=12.0

# Environment management
python-dotenv>=1.0.0

# Terminal formatting
rich>=13.0.0

# Async file I/O
aiofiles>=23.0.0
